requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.19
reportlab==4.2.5
openpyxl==3.1.5
python-dateutil==2.9.0
email-validator==2.2.0
pytesseract==0.3.13
Pillow==11.0.0
pdf2image==1.17.0
